Replaced some static_casts with NumericCast() and ToUnderlying() 04/16/2023 | 23w15b1

+++ TRAP/src/Graphics/API/ShaderReflection.h
@@ -60,7 +60,7 @@ namespace TRAP::Graphics::API::ShaderReflection
 	struct ShaderVariable
 	{
 		//Parents resource index
-		uint32_t ParentIndex{};
+		uint64_t ParentIndex{};
 
 		//The offset of the Variable.
 		uint32_t Offset{};
